Change return value for some functions to reflect the ptrace signature.

+++ src/roach.h
@@ -78,11 +78,11 @@ int roach_wait (roach_context_t *ctx, int *status);
 bool roach_entering_sc_p (roach_context_t *ctx);
 
 /* Architecture specific functions.  */
-int roach_get_sc (roach_context_t *ctx);
-int roach_set_sc (roach_context_t *ctx, int syscall);
-int roach_set_sc_ret (roach_context_t *ctx, int retval);
-int roach_get_sc_arg (roach_context_t *ctx, int arg);
-int roach_set_sc_arg (roach_context_t *ctx, int arg, void *data);
+long roach_get_sc (roach_context_t *ctx);
+long roach_set_sc (roach_context_t *ctx, int syscall);
+long roach_set_sc_ret (roach_context_t *ctx, int retval);
+long roach_get_sc_arg (roach_context_t *ctx, int arg);
+long roach_set_sc_arg (roach_context_t *ctx, int arg, void *data);
 int roach_write_mem (roach_context_t *ctx, const char const *data,
                      const char const *addr, size_t len);
 int roach_read_mem (roach_context_t *ctx, char *data,
